requirements.txt
dash==2.14.2
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
databricks-sdk==0.19.0
databricks-sql-connector==2.9.3
dash-bootstrap-components==1.5.0 